proof mode: meta code > TC inference
This is a large refactor of the design of the Iris proof mode to use
regular theorems instead of kernel reduction and typeclass inference
to perform the main logic.

In the new design, proof goals are represented by a `?m : Entails' P Q`
metavariable, where `P` is a tree of the form:

* `emp`
* `P1 ∗ P2` where `P1` and `P2` are trees
* `[mdata name: HA] A`, representing a spatial assumption
  displayed as `∗HA : A`
* `□ [mdata name: HA] A`, representing an intuitionistic assumption
  displayed as `□HA : A`

The tree is not assumed to have any particular structure, but it will
typically be a left-leaning list.

This type contains the variable names in mdata nodes, but we try to avoid
these getting into the proof term, and the mdata is stripped from the
generated proof term using `Hyps.strip`. One consequence of this
representation compared to the previous / Coq's "two lists" approach
is that intuitionistic assumptions are no longer guaranteed to precede
spatial assumptions, which is why they are denoted using sigils on each
variable instead of having two big separators.

The main logic of tactics like cases now walk over the hypothesis tree
in meta code, using `Qq` heavily to help ensure type correctness of
generated expressions. That means we don't need to reimplement `List`
with all definitions `@[reducible]`, and moreover we get much saner proof
terms which amount to calling theorems from `DerivedLaws.lean`.
It is also measurably faster:
`lake env time lean src/Iris/Tests/Tactics.lean` has decreased from
4.69 secs to 1.46 secs as a result of this commit.

Other assorted changes:

* The `Intuitionistic P` class, meaning `P ⊢ □ P`, was added to support
  one place where the original proof used `Δ.spatial.isEmpty` to justify
  adding `□` on the context; in the new representation this is best done
  by a typeclass which observes that all the leaves have `□ P`. This
  follows from some other classes but there wasn't any single class which
  did what we want here.

* The proof mode display has changed, from:
  ```
  Iris Proof Mode
  ─────────────────────────────────────
  HP : P
  ───────────────────────────────────── □
  HQ : Q
  HR : R
  ───────────────────────────────────── ∗
  Q
  ```
  to
  ```
  □HP : P
  ∗HQ : Q
  ∗HR : R
  ⊢ Q
  ```
  which blends in a bit better with the lean context. (Fingers crossed
  that we can get fancy HTML support, so that we can color the variables
  differently to help them stand out more.)

* The tactics have been moved to separate files, to help with build
  parallelism. Theorems have also been placed next to the tactics that
  use them rather than gathered in a separate file.

* The syntax of cases patterns has changed, to use `-` instead of `_`
  for clearing hypotheses (for consistency with `rcases`, and also
  because `_` elsewhere in lean means to introduce a variable with a
  hygienic name, not to remove the variable).

* The cases pattern syntax for "move to spatial context" has changed from
  `-□pat` (which I guess is meant to suggest "not `□`") to `∗pat`,
  mimicking the use of prefix-`∗` for spatial hypotheses in the display.

* The syntax for rename has changed from `irename HA to HB` to
  `irename HA => HB` for consistency with lean core, although the core
  tactic `rename T => h` actually means to find the first hypothesis with
  type `T` and rename it to `h`.

* `iintro x` on a forall statement works directly rather than requiring
  `iintro ⌜x⌝`. I don't think there is any need for this, as the variable
  is already a pure lean value, not a ⌜φ⌝ expression.

* `iassumption` no longer attempts to close the goal if it finds `False`
  in the context, for consistency with the lean tactic. Instead you can
  use the `⟨⟩` pattern on a variable of type `False`, or use `iexfalso`.

* `ispecialize` takes terms, not just variables, so you can write
  `ispecialize H (x + 1)` to instantiate a forall hypothesis. The `as H`
  part is now optional (for consistency with the lean tactic) and means
  to modify the existing hypothesis.