Add ProofMode.SynthInstance
Add a custom version of synthInstance for the proof mode that can
create and instantiate mvars. It is implemented as a simple
backtracking search, reusing the standard instance infrastructure. This
version of synthInstance should be used for the IPM proof mode
classes, marked with the attribute ipm_class.

Also some other small refactorings:
- Allow iexists to take multiple arguments
- Fix handling of created goals in iexists
- Use the Goals structure more consistently
- Add some more error messages
- Generalize AsEmpValid over the direction similar to the Rocq version