Add InOut for dynamic input / output parameters of proof mode classes
This commit adds [InOut] that is used to dynamically determine whether
a type class parameter is an input or an output. This is important for
classes that are used with multiple modings, e.g., IntoWand. Instances
can match on the InOut parameter to avoid accidentially instantiating
outputs if matching on an input was intended.